LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/input-whitespace-expected.txt



FAIL whitespace should not be eaten after parent style recalculation assert_equals: expected "a b" but got ""
